+++ astroquery/mast/tests/test_mast.py
@@ -6,6 +6,7 @@ import re
 import warnings
 from shutil import copyfile
 from unittest.mock import patch
+from pathlib import Path
 
 import pytest
 import numpy as np
@@ -13,6 +14,7 @@ import numpy as np
 from astropy.table import Table, unique
 from astropy.coordinates import SkyCoord
 from astropy.io import fits
+from botocore.exceptions import ClientError
 
 import astropy.units as u
 from requests import HTTPError, Response
@@ -53,7 +55,7 @@ DATA_FILES = {'Mast.Caom.Cone': 'caom.json',
               'Mast.HscMatches.Db.v3': 'matchid.json',
               'Mast.HscMatches.Db.v2': 'matchid.json',
               'Mast.HscSpectra.Db.All': 'spectra.json',
-              'mast_relative_path': 'mast_relative_path.json',
+              'get_cloud_paths': 'mast_relative_path.json',
               'panstarrs': 'panstarrs.json',
               'panstarrs_columns': 'panstarrs_columns.json',
               'tess_cutout': 'astrocut_107.27_-70.0_5x5.zip',
@@ -78,7 +80,6 @@ def patch_post(request):
     mp.setattr(mast.auth.MastAuth, 'session_info', session_info_mockreturn)
 
     mp.setattr(mast.Observations, '_download_file', download_mockreturn)
-    mp.setattr(mast.Observations, 'download_file', download_mockreturn)
     mp.setattr(mast.Catalogs, '_download_file', download_mockreturn)
     mp.setattr(mast.MastMissions, '_download_file', download_mockreturn)
     mp.setattr(mast.Tesscut, '_download_file', tesscut_download_mockreturn)
@@ -149,7 +150,7 @@ def request_mockreturn(url, params={}):
     elif 'panstarrs' in url:
         filename = data_path(DATA_FILES['panstarrs_columns'])
     elif 'path_lookup' in url:
-        filename = data_path(DATA_FILES['mast_relative_path'])
+        filename = data_path(DATA_FILES['get_cloud_paths'])
     with open(filename, 'rb') as infile:
         content = infile.read()
     return MockResponse(content)
@@ -894,15 +895,174 @@ def test_observations_download_products(patch_post, tmpdir):
     result1 = mast.Observations.download_products(products[0], download_dir=str(tmpdir))
     assert isinstance(result1, Table)
 
+    # Warn if no products to download
+    with pytest.warns(NoResultsWarning, match='No products to download'):
+        result = mast.Observations.download_products('2003738726',
+                                                     download_dir=str(tmpdir),
+                                                     productType=["INVALID_TYPE"])
+        assert result is None
 
-def test_observations_download_file(patch_post, tmpdir):
-    # pull a single data product
-    products = mast.Observations.get_product_list('2003738726')
-    uri = products['dataURI'][0]
+    # Warn if curl_flag and flags are both set
+    with pytest.warns(InputWarning, match='flat=True has no effect on curl downloads.'):
+        result = mast.Observations.download_products('2003738726',
+                                                     curl_flag=True,
+                                                     flat=True)
+        assert isinstance(result, Table)
 
-    # download it
-    result = mast.Observations.download_file(uri)
+
+@patch('boto3.resource')
+@patch('boto3.client')
+@patch.object(Path, "is_file", return_value=True)
+def test_observations_download_products_cloud(mock_is_file, mock_client, mock_resource, patch_post,
+                                              monkeypatch):
+    pytest.importorskip("boto3")
+    mock_client.return_value.head_object.return_value = {'ContentLength': 12345}
+    mock_resource.return_value.Bucket.return_value.download_file.return_value = None
+    obsid = '2003738726'
+    data_uri = 'mast:HST/product/u9o40504m_c3m.fits'
+
+    # Enable access to public AWS S3 bucket
+    mast.Observations.enable_cloud_dataset()
+
+    result = mast.Observations.download_products(obsid,
+                                                 dataURI=data_uri)
+    assert isinstance(result, Table)
+    assert result[0]['Status'] == 'COMPLETE'
+
+    # Mock cloud download failure, fallback to on-prem
+    mock_resource.return_value.Bucket.return_value.download_file.side_effect = Exception('Not found in cloud')
+    # Check that info message is logged
+    with pytest.warns(InputWarning, match='Falling back to MAST download'):
+        result = mast.Observations.download_products(obsid,
+                                                     dataURI=data_uri)
+    assert result[0]['Status'] == 'COMPLETE'
+
+    # Cloud download failure, do not fallback to on-prem
+    with pytest.warns(NoResultsWarning, match='Skipping download.'):
+        result = mast.Observations.download_products(obsid,
+                                                     dataURI=data_uri,
+                                                     cloud_only=True)
+    assert result[0]['Status'] == 'SKIPPED'
+
+    # Products not found in cloud, skip download
+    monkeypatch.setattr(mast.Observations, 'get_cloud_uris', lambda *a, **k: {})
+    with pytest.warns(NoResultsWarning, match='was not found in the cloud. Skipping download.'):
+        result = mast.Observations.download_products(obsid,
+                                                     dataURI=data_uri,
+                                                     cloud_only=True)
+    assert result[0]['Status'] == 'SKIPPED'
+    assert result[0]['Message'] == 'Product not found in cloud'
+
+    # Products not found in cloud, fall back
+    with pytest.warns(InputWarning, match='was not found in the cloud. Falling back to MAST download'):
+        result = mast.Observations.download_products(obsid,
+                                                     dataURI=data_uri)
+    assert result[0]['Status'] == 'COMPLETE'
+
+    mast.Observations.disable_cloud_dataset()
+
+    # Cloud access not enabled, warn if cloud_only is True
+    with pytest.warns(InputWarning, match='cloud data access is not enabled'):
+        result = mast.Observations.download_products('2003738726',
+                                                     dataURI='mast:HST/product/u9o40504m_c3m.fits',
+                                                     cloud_only=True)
+    assert result[0]['Status'] == 'COMPLETE'
+
+
+@patch.object(Path, "is_file", return_value=True)
+def test_observations_download_file(mock_is_file, patch_post, tmpdir):
+    mast_uri = 'mast:HST/product/u9o40504m_c3m.fits'
+
+    result = mast.Observations.download_file(mast_uri, local_path=tmpdir)
     assert result == ('COMPLETE', None, None)
+
+    unauth_uri = 'mast:HST/product/unauthorized.fits'
+    result = mast.Observations.download_file(unauth_uri)
+    assert result[0] == 'ERROR'
+    assert 'HTTPError' in result[1]
+
+
+def test_observations_download_file_not_found(patch_post, tmpdir):
+    mast_uri = 'mast:HST/product/u9o40504m_c3m.fits'
+
+    result = mast.Observations.download_file(mast_uri, local_path=tmpdir)
+    assert result[0] == 'ERROR'
+    assert result[1] == 'File was not downloaded'
+
+
+@patch('boto3.resource')
+@patch('boto3.client')
+@patch.object(Path, "is_file", return_value=True)
+def test_observations_download_file_cloud(mock_is_file, mock_client, mock_resource, patch_post):
+    pytest.importorskip("boto3")
+    mock_client.return_value.head_object.return_value = {'ContentLength': 12345}
+    mock_resource.return_value.Bucket.return_value.download_file.return_value = None
+
+    # Enable access to public AWS S3 bucket
+    mast.Observations.enable_cloud_dataset()
+    mast_uri = 'mast:HST/product/u9o40504m_c3m.fits'
+
+    # Warn if both cloud_only and force_on_prem are True
+    with pytest.warns(InputWarning, match='Both `cloud_only` and `force_on_prem` are set to True'):
+        result = mast.Observations.download_file(mast_uri, cloud_only=True, force_on_prem=True)
+        assert result == ('COMPLETE', None, None)
+
+    # Skip file if cloud_only is True but file is not in cloud
+    nonexistent_uri = 'mast:HST/product/does_not_exist.fits'
+    with pytest.warns(NoResultsWarning, match=f'The product {nonexistent_uri} was not found in the cloud'):
+        result = mast.Observations.download_file(nonexistent_uri, cloud_only=True)
+        assert result == ('SKIPPED', None, None)
+
+    # Use on-prem download if cloud_only is False and file is not in cloud
+    with pytest.warns(InputWarning, match=f'The product {nonexistent_uri} was not found in the cloud'):
+        result = mast.Observations.download_file(nonexistent_uri, cloud_only=False)
+        assert result == ('COMPLETE', None, None)
+
+    # Error if data product is not a string
+    with pytest.raises(InvalidQueryError, match='A valid data product URI'):
+        mast.Observations.download_file(12345, cloud_only=True)
+
+    mast.Observations.disable_cloud_dataset()
+
+    # Warning if cloud dataset is not enabled
+    with pytest.warns(InputWarning, match='cloud data access is not enabled'):
+        result = mast.Observations.download_file(mast_uri, cloud_only=True)
+        assert result == ('COMPLETE', None, None)
+
+
+@patch('boto3.client')
+def test_observations_list_cloud_missions(mock_client, patch_post):
+    pytest.importorskip('boto3')
+    mock_client.return_value.list_objects_v2.return_value = {
+        'CommonPrefixes': [{'Prefix': 'hst/'}, {'Prefix': 'jwst/'}, {'Prefix': 'mast/'}]
+    }
+
+    mast.Observations.enable_cloud_dataset()
+    supported = mast.Observations.list_cloud_datasets()
+    assert isinstance(supported, list)
+    assert 'hst' in supported
+    assert 'jwst' in supported
+    assert 'mast' in supported
+
+    mast.Observations.disable_cloud_dataset()
+
+
+@patch('boto3.client')
+def test_observations_list_cloud_missions_error(mock_client, patch_post, caplog):
+    pytest.importorskip('boto3')
+
+    # Error without cloud connection
+    with pytest.raises(RemoteServiceError):
+        mast.Observations.list_cloud_datasets()
+
+    # Mock an error when listing objects
+    mock_client.return_value.list_objects_v2.side_effect = Exception('AWS error')
+
+    mast.Observations.enable_cloud_dataset()
+    supported = mast.Observations.list_cloud_datasets()
+    assert supported == []
+
+    mast.Observations.disable_cloud_dataset()
 
 
 @patch('boto3.client')
@@ -929,6 +1089,10 @@ def test_observations_get_cloud_uri(mock_client, patch_post):
     # String input
     uri = mast.Observations.get_cloud_uri(mast_uri)
     assert uri == expected
+
+    # Warn if not found
+    with pytest.warns(NoResultsWarning, match='Failed to retrieve cloud path'):
+        mast.Observations.get_cloud_uri('mast:HST/product/does_not_exist.fits')
 
     mast.Observations.disable_cloud_dataset()
 
@@ -962,6 +1126,11 @@ def test_observations_get_cloud_uris(mock_client, patch_post):
     assert len(uris) == 1
     assert uris[0] == expected
 
+    # Return full URLs
+    uris = mast.Observations.get_cloud_uris([mast_uri], include_bucket=False, full_url=True)
+    assert isinstance(uris, list)
+    assert uris[0].startswith('http://s3.amazonaws.com/')
+
     # Return a map of URIs
     uri_map = mast.Observations.get_cloud_uris([mast_uri], return_uri_map=True)
     assert isinstance(uri_map, dict)
@@ -974,8 +1143,34 @@ def test_observations_get_cloud_uris(mock_client, patch_post):
                                          extension='png')
 
     # Warn if not found
-    with pytest.warns(NoResultsWarning, match='Failed to retrieve MAST relative path'):
+    with pytest.warns(NoResultsWarning, match='Failed to retrieve cloud path'):
         mast.Observations.get_cloud_uris(['mast:HST/product/does_not_exist.fits'])
+
+    mast.Observations.disable_cloud_dataset()
+
+
+@patch('boto3.client')
+def test_observations_get_cloud_uris_error(mock_client, patch_post):
+    pytest.importorskip("boto3")
+
+    # Mock head_object to raise an exception
+    # Raise the error if not a 404
+    exc = ClientError({'Error': {'Code': '500', 'Message': 'Internal Server Error'}}, 'HeadObject')
+    mock_client.return_value.head_object.side_effect = exc
+
+    mast.Observations.enable_cloud_dataset()
+    with pytest.raises(ClientError):
+        mast.Observations.get_cloud_uris(['mast:HST/product/u9o40504m_c3m.fits'])
+
+    # Only warn if the error is a 404
+    exc = ClientError({'Error': {'Code': '404', 'Message': 'Not Found'}}, 'HeadObject')
+    mock_client.return_value.head_object.side_effect = exc
+
+    with pytest.warns(NoResultsWarning, match='Failed to retrieve cloud path'):
+        uris = mast.Observations.get_cloud_uris(['mast:HST/product/u9o40504m_c3m.fits'])
+    assert uris == []
+
+    mast.Observations.disable_cloud_dataset()
 
 
 @patch('boto3.client')
@@ -998,6 +1193,8 @@ def test_observations_get_cloud_uris_query(mock_client, patch_post):
     with pytest.warns(NoResultsWarning, match='No matching products'):
         mast.Observations.get_cloud_uris(target_name=234295610,
                                          filter_products={'productSubGroupDescription': 'LC'})
+
+    mast.Observations.disable_cloud_dataset()
 
 
 ######################
@@ -1485,3 +1682,92 @@ def test_parse_input_location(patch_post):
     with pytest.warns(InputWarning, match="Resolver is only used when resolving object names"):
         loc = mast.utils.parse_input_location(coordinates=coord, resolver="SIMBAD")
         assert isinstance(loc, SkyCoord)
+
+
+################
+# Cloud tests #
+################
+
+@patch("boto3.resource")
+@patch("boto3.client")
+def test_download_file_from_cloud(mock_client, mock_resource, patch_post):
+    pytest.importorskip("boto3")
+
+    cloud = mast.cloud.CloudAccess()
+
+    mock_client.return_value.head_object.return_value = {'ContentLength': 123}
+    mock_resource.return_value.Bucket.return_value.download_file.return_value = None
+
+    cloud.download_file_from_cloud(
+        "s3://stpubdata/hst/public/u9o4/u9o40504m/u9o40504m_c3m.fits",
+        "local.fits",
+        verbose=False
+    )
+    mock_resource.return_value.Bucket.return_value.download_file.assert_called_once()
+
+
+@patch("boto3.resource")
+@patch("boto3.client")
+def test_download_file_from_cloud_not_found(mock_client, mock_resource, patch_post):
+    pytest.importorskip("boto3")
+
+    cloud = mast.cloud.CloudAccess()
+
+    # Force get_cloud_uri_list to return [None]
+    cloud.get_cloud_uri_list = lambda *a, **k: [None]
+
+    with pytest.raises(RemoteServiceError, match='was not found in the cloud'):
+        cloud.download_file_from_cloud(
+            "mast:HST/product/missing.fits",
+            "local.fits",
+        )
+
+
+@patch('os.path.exists', return_value=True)
+@patch('os.path.getsize', return_value=123)
+@patch('boto3.resource')
+@patch('boto3.client')
+def test_download_file_from_cloud_existing(mock_client, mock_resource, mock_getsize, mock_exists, patch_post):
+    pytest.importorskip("boto3")
+
+    mock_client.return_value.head_object.return_value = {'ContentLength': 123}
+    cloud = mast.cloud.CloudAccess()
+
+    # File exists locally with same size
+    cloud.download_file_from_cloud(
+        "mast:HST/product/u9o40504m_c3m.fits",
+        "local.fits",
+        verbose=False
+    )
+    # No download should be attempted
+    mock_resource.return_value.Bucket.return_value.download_file.assert_not_called()
+
+    # File exists locally with different size
+    mock_getsize.return_value = 456
+    cloud.download_file_from_cloud(
+        "mast:HST/product/u9o40504m_c3m.fits",
+        "local.fits",
+        verbose=False
+    )
+    # Download should be attempted
+    mock_resource.return_value.Bucket.return_value.download_file.assert_called_once()
+
+
+@patch("boto3.resource")
+@patch("boto3.client")
+def test_download_file_from_cloud_verbose(mock_client, mock_resource, patch_post):
+    pytest.importorskip("boto3")
+
+    cloud = mast.cloud.CloudAccess()
+
+    mock_client.return_value.head_object.return_value = {'ContentLength': 123}
+    mock_resource.return_value.Bucket.return_value.download_file.return_value = None
+
+    cloud.download_file_from_cloud(
+        "s3://stpubdata/hst/public/u9o4/u9o40504m/u9o40504m_c3m.fits",
+        "local.fits",
+        verbose=True
+    )
+    # Ensure callback was supplied
+    _, kwargs = mock_resource.return_value.Bucket.return_value.download_file.call_args
+    assert "Callback" in kwargs